Organization, cleanup, pydantic serialization, update tests
